+++ examples/expo-family-shopping-list/src/app/(auth)/sign_up.tsx
@@ -1,11 +1,9 @@
 import React, { useState } from 'react'
 import { Alert, View } from 'react-native'
 import { Button } from 'react-native-paper'
-import EmailPasswordForm from '../components/EmailPasswordForm'
+import EmailPasswordForm from '../../components/EmailPasswordForm'
 import { Link, router, useLocalSearchParams } from 'expo-router'
-import { useAuthActions } from '../components/AuthProvider'
-import AppLogo from '../components/AppLogo'
-import { SafeAreaView } from 'react-native-safe-area-context'
+import { useAuthActions } from '../../components/AuthProvider'
 
 export default function SignUp() {
   const { email: initialEmail } = useLocalSearchParams<{ email: string }>()
@@ -21,10 +19,7 @@ export default function SignUp() {
   }
 
   return (
-    <SafeAreaView style={{ margin: 16, gap: 12 }}>
-      <View style={{ alignItems: 'center' }}>
-        <AppLogo height={200} width={200} />
-      </View>
+    <View style={{ gap: 12 }}>
       <EmailPasswordForm
         initialEmail={initialEmail}
         passwordPlaceholder='Must have at least 6 characters'
@@ -37,6 +32,6 @@ export default function SignUp() {
           I already have an account
         </Button>
       </Link>
-    </SafeAreaView>
+    </View>
   )
 }
